# schema/rest/openapi.yaml
openapi: 3.0.3
info:
  title: Neuron Visualizer API
  version: 0.0.1
  description: |
    This is a OpenAPI specification for Neuron Visualizer. It is used to generate server(golang) and client(unity) code.
  license:
    name: MIT License
    url: https://spdx.org/licenses/MIT.html
servers:
  - url: http://localhost:8080
    description: Local Server (Default)
tags:
  - name: Health
    description: Health Check API
  - name: Neuron
    description: Neuron API
paths:
  /api/v1/health:
    get:
      tags:
        - Health
      summary: Health check
      description: Health check API
      operationId: healthCheck
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthCheckResponse'
  /api/v1/neurons:
    get:
      tags:
        - Neuron
      summary: Get neurons available for visualization
      description: Get neurons available for visualization
      operationId: getNeurons
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetNeuronsResponse'
  /api/v1/neurons/{neuron_name}/compartments:
    get:
      tags:
        - Neuron
      summary: Get compartments of the neuron
      description: Get compartments of the neuron
      operationId: getNeuronCompartments
      parameters:
        - name: neuron_name
          in: path
          description: Name of the neuron
          required: true
          schema:
            $ref: '#/components/schemas/NeuronName'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetNeuronCompartmentsResponse'
  /api/v1/neurons/{neuron_name}/compartments/membranePotentials:
    get:
      tags:
        - Neuron
      summary: Get membrane potentials of the neuron
      description: Get membrane potentials of the neuron
      operationId: getNeuronMembranePotentials
      parameters:
        - name: neuron_name
          in: path
          description: Name of the neuron
          required: true
          schema:
            $ref: '#/components/schemas/NeuronName'
      responses:
        '200':
          description: OK
          content:
            text/event-stream:
              schema:
                type: array
                format: event-stream
                items:
                  $ref: '#/components/schemas/GetNeuronCompartmentsMembranePotentialResponse'
components:
  schemas:
    HealthCheckResponse:
      type: object
      properties:
        status:
          type: string
          description: Status of the health check
          example: OK
          enum:
            - OK
            - ERROR
        message:
          type: string
          description: Message of the health check
          example: server is running
      required:
        - status
        - message
    GetNeuronsResponse:
      type: object
      properties:
        name:
          $ref: '#/components/schemas/NeuronName'
      required:
        - name
    GetNeuronCompartmentsResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          minimum: 0
          description: ID of the compartment, unique for the same neuron.
          example: 0
        type:
          $ref: '#/components/schemas/NeuronCompartmentType'
        positionX:
          $ref: '#/components/schemas/Coordinate'
        positionY:
          $ref: '#/components/schemas/Coordinate'
        positionZ:
          $ref: '#/components/schemas/Coordinate'
        radius:
          type: number
          format: double
          minimum: 0
          description: Radius of the compartment.
          example: 5.4428
        parentID:
          type: integer
          format: int64
          minimum: -1
          description: ID of the parent compartment. -1 indicates no parent compartment.
          example: -1
      required:
        - id
        - type
        - positionX
        - positionY
        - positionZ
        - radius
        - parentID
    GetNeuronCompartmentsMembranePotentialResponse:
      type: object
      properties:
        timeStep:
          type: number
          format: double
          description: Step Time of the membrane potential.
          example: 0
        membranePotentials:
          type: array
          items:
            type: number
            format: double
            description: Membrane potential of the compartment.
            example: -65.0
      required:
        - timeStep
        - membranePotentials
    NeuronName:
      type: string
      description: Name of the neuron
      example: cerebral_cortex_pyramidal_cell
    NeuronCompartmentType:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID of the compartment type.
          example: 1
        name:
          type: string
          description: |
            Type of the compartment.
              SOMA (細胞体).
              AXON (軸索).
              BASAL_DENDRITE (基底樹状突起).
              APICAL_DENDRITE (尖端樹状突起).
          example: SOMA
          enum:
            - SOMA
            - AXON
            - BASAL_DENDRITE
            - APICAL_DENDRITE
      required:
        - id
        - name
    Coordinate:
      type: number
      format: double
      description: Coordinate for example x, y, z
      example: 379.4648